src/requirements.txt
python-dotenv
slack_bolt